sophgo rebuilt tdl_sdk
commit dfc3d2ebc12ef13e82e59b9a4224f9a66be22ecc
Author: zhenxing.ye <zhenxing.ye@sophgo.com>
Date:   Thu Dec 26 18:08:03 2024 +0800

    update readme

+++ tdl_sdk/modules/core/ocr/ocr_detection/ocr_detection.cpp
@@ -121,8 +121,9 @@ void OCRDetection::outputParser(float thresh, float boxThresh, cvtdl_object_t *o
 
     cv::Rect boundingBox = cv::boundingRect(contour);
 
-    cvtdl_bbox_t bbox = {boundingBox.x, boundingBox.y, boundingBox.x + boundingBox.width,
-                         boundingBox.y + boundingBox.height};
+    cvtdl_bbox_t bbox = {static_cast<float>(boundingBox.x), static_cast<float>(boundingBox.y),
+                         static_cast<float>(boundingBox.x + boundingBox.width),
+                         static_cast<float>(boundingBox.y + boundingBox.height)};
     cvtdl_bbox_t rescaled_bbox =
         box_rescale(obj_meta->width, obj_meta->height, shape.dim[3], shape.dim[2], bbox,
                     meta_rescale_type_e::RESCALE_CENTER);
